+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Security in case of vulnerabilities.
 - Added `Dao Revocation` workflow.
 - Configured Cache settings.
 - Added `Confirm the date the academy opened` task.
+- Added 'Closing of transfer or conversion project'.
 - Added `Church supplemental agreement` task for both conversion and transfer projects.
 
 ### Changed
@@ -29,6 +30,7 @@ Security in case of vulnerabilities.
 ### Fixed
 
 - Fixed broken task note urls.
+- Fixed all project by month query for conversion projects if all conditions met is null
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-09-25.771...main) for everything awaiting release